improvement: merge quick add logic for all layouts. (#5323)
