+++ web/ce/components/issues/index.ts
@@ -2,3 +2,4 @@ export * from "./bulk-operations";
 export * from "./worklog";
 export * from "./issue-modal";
 export * from "./issue-details";
+export * from "./quick-add";
