+++ web/core/components/issues/issue-layouts/gantt/index.ts
@@ -1,3 +1,3 @@
 export * from "./blocks";
 export * from "./base-gantt-root";
-export * from "./quick-add-issue-form";
+
