+++ web/core/components/issues/issue-layouts/index.ts
@@ -18,3 +18,7 @@ export * from "./properties";
 
 // save view
 export * from "./save-filter-view";
+
+// quick add
+export * from "./quick-add";
+
