+++ web/core/components/issues/issue-layouts/kanban/index.ts
@@ -1,4 +1,3 @@
 export * from "./block";
 export * from "./roots";
 export * from "./blocks-list";
-export * from "./quick-add-issue-form";
