+++ web/core/components/issues/issue-layouts/list/index.ts
@@ -3,4 +3,3 @@ export * from "./block-root";
 export * from "./block";
 export * from "./roots";
 export * from "./blocks-list";
-export * from "./quick-add-issue-form";
