+++ web/core/components/issues/issue-layouts/quick-add/button/index.ts
@@ -1,0 +1,4 @@
+export * from "./list";
+export * from "./kanban";
+export * from "./gantt";
+export * from "./spreadsheet";
