+++ web/core/components/issues/issue-layouts/quick-add/form/index.ts
@@ -1,0 +1,5 @@
+export * from "./list";
+export * from "./kanban";
+export * from "./gantt";
+export * from "./calendar";
+export * from "./spreadsheet";
