+++ web/core/components/issues/issue-layouts/quick-add/index.ts
@@ -1,0 +1,3 @@
+export * from "./root";
+export * from "./form";
+export * from "./button";
