+++ web/ee/components/issues/quick-add/index.ts
@@ -1,0 +1,1 @@
+export * from "./root";
