+++ web/ee/components/issues/quick-add/root.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/issues/quick-add/root";
